+++ src/FR_math.h
@@ -32,8 +32,13 @@
 #ifndef __FR_Math_h__
 #define __FR_Math_h__
 
-#define FR_MATH_VERSION     "2.0.6"
-#define FR_MATH_VERSION_HEX  0x020006  /* major << 16 | minor << 8 | patch */
+#define FR_MATH_VERSION     "2.0.7"
+#define FR_MATH_VERSION_HEX  0x020007  /* major << 16 | minor << 8 | patch */
+
+#ifdef FR_CORE_ONLY
+#define FR_NO_PRINT
+#define FR_NO_WAVES
+#endif
 
 #ifdef __cplusplus
 extern "C"